# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Docs for LinkVault API",
        "title": "LinkVault API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/folders": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "folder"
                ],
                "summary": "Get many folders belongs to user",
                "parameters": [
                    {
                        "type": "string",
                        "default": "updatedAt_DESC",
                        "description": "Order by title, created date, or modified date",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit every page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page count",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Folder(s) found",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-entities_FolderEntity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "folder"
                ],
                "summary": "Create a folder",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateFolderDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created a folder",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-entities_FolderEntity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-any"
                        }
                    }
                }
            }
        },
        "/folders/{folderId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "folder"
                ],
                "summary": "Get details about a folder belongs to user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder id",
                        "name": "folderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Folder(s) found",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-entities_FolderEntity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "folder"
                ],
                "summary": "Delete a folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "folderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Folder deleted",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-entities_FolderEntity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-any"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "folder"
                ],
                "summary": "Update folder detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "folderId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update folder payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateFolderDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Folder(s) found",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-entities_FolderEntity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.CreateFolderDto": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dtos.UpdateFolderDto": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "entities.FolderEntity": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
        "utils.BaseResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.BaseResponse-entities_FolderEntity": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entities.FolderEntity"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}